refactor: extract catalog introspection to own crate
Extracts src/services/catalog_introspection.rs (1,902 LOC) to a new
proximadb-catalog-introspection crate under crates/binding/. The root crate
keeps a re-export shim so all callers (protocol.rs, handlers.rs, services/mod.rs)
resolve unchanged via crate::services::CatalogIntrospectionService.

Dependency wiring (avoids a cyclic dep):
- New crate depends on proximadb-catalog (sub-crate) for catalog types +
  sqlparser for the parse_select_aliases helper (inlined, 35 lines, kept in
  lockstep with src/query/sql_frontend/parser.rs) — NO production dep on root.
- Root depends on the new crate (runtime, for the shim).
- New crate's integration tests use root as a dev-dep (DdlService/SqlFrontendParser);
  dev-deps do not create build cycles, mirroring the proximadb-embedded precedent.

Adds the crate to the CI feature-matrix so cfg drift is caught at every tier.
