+++ crates/binding/proximadb-catalog-introspection/Cargo.toml
@@ -1,0 +1,38 @@
+[package]
+name = "proximadb-catalog-introspection"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "Catalog introspection service for ProximaDB — projects the native xCatalog into SQL-compatible metadata tables. Extracted from the root crate."
+
+[lib]
+name = "proximadb_catalog_introspection"
+path = "src/lib.rs"
+
+[dependencies]
+# Depends on the extracted catalog crate (NOT the root proximadb crate) to avoid a
+# cyclic dependency — the root crate re-exports this crate via a shim, so this
+# crate must not depend back on root in production code. Catalog types
+# (CatalogManager, CatalogTableSchema, TableIdentifier, …) live in proximadb-catalog.
+proximadb-catalog = { path = "../../control/proximadb-catalog" }
+proximadb-data-model = { path = "../../foundation/proximadb-data-model" }
+anyhow.workspace = true
+serde.workspace = true
+# parse_select_aliases is a thin wrapper over sqlparser-rs (same version the root
+# sql_frontend uses). Inlined here to keep this crate free of a root-crate dep.
+sqlparser = { version = "0.59", features = ["visitor"] }
+
+[dev-dependencies]
+# Integration tests exercise the service through the root crate's DDL/frontend
+# surfaces. Dev-deps do not create build cycles (test-mode only) — mirrors the
+# proximadb-embedded ↔ root precedent.
+proximadb = { path = "../../.." }
+tempfile.workspace = true
+tokio.workspace = true
+
+[lints]
+workspace = true
